chore: split clickhouse tests into multiple files
Ensures they run in CI as separate jobs
